Django_01(reservation)/pyproject.toml
[project]
name = "django-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
]
